fix: don't overwrite config file
So far we have written a default config file if we encountered any error
when trying to read the one that should exist. That included:
* missing key-value-pairs,
* empty file, and
* a non-existant file.

Therefore, when a field is added to the Config struct, the entire config
file was overwritten with a default config. This occured during #20.

Instead, we now provide default values in form of generator functions to
serde. Any value that is present in the config is then parsed and
replaces the default value. Only if the config file does not exist we
create an empty one.

Closes #26.