fix(rustup): Account for version date
The previous implementation skipped the date of
the release in case the nightly was not the latest
one. This commit relies on the fact that the
rustup toolchains have a target triple at the end
of the toolchain name when listed and is prone to
break/error should that not be true for all
toolchains. This covers, in particular, symlinks
for any custom toolchains.

Signed-off-by: innocentzero <isfarulhaque@proton.me>